# openapi.yaml
openapi: 3.0.0
paths:
  /health:
    get:
      operationId: HealthController_getHealth
      parameters: []
      responses:
        "200":
          description: Health status response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponseDto"
      summary: Check API service and database health status
      tags:
        - Health
  /auth/register:
    post:
      operationId: AuthController_register
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      summary: Register a new user account
      tags: &a1
        - Auth
  /auth/login:
    post:
      operationId: AuthController_login
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponseDto"
      summary: Authenticate user with email and password
      tags: *a1
  /auth/refresh:
    post:
      operationId: AuthController_refresh
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshResponseDto"
      summary: Rotate refresh token and get a new access token
      tags: *a1
  /auth/logout:
    post:
      operationId: AuthController_logout
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutDto"
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      summary: Logout and revoke refresh token session
      tags: *a1
  /auth/me:
    get:
      operationId: AuthController_me
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentUserProfileDto"
      security:
        - bearer: []
      summary: Get current authenticated user profile
      tags: *a1
  /devices:
    post:
      operationId: DeviceController_linkDevice
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkDeviceDto"
      responses:
        "200":
          description: Existing installation updated and new device token rotated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDeviceResponseDto"
        "201":
          description: Device linked for the first time
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkDeviceResponseDto"
      security:
        - bearer: []
      summary: Link or re-link a mobile device to the authenticated user account
      tags: &a2
        - Devices
    get:
      operationId: DeviceController_listDevices
      parameters: []
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListDevicesResponseDto"
      security:
        - bearer: []
      summary: List all devices belonging to the authenticated user (newest first)
      tags: *a2
  /devices/{id}:
    get:
      operationId: DeviceController_getDevice
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponseDto"
      security:
        - bearer: []
      summary: Get details of a specific device belonging to the user
      tags: *a2
    patch:
      operationId: DeviceController_updateDevice
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponseDto"
      security:
        - bearer: []
      summary: Update device details (name, fcmToken)
      tags: *a2
    delete:
      operationId: DeviceController_unlinkDevice
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "204":
          description: ""
      security:
        - bearer: []
      summary: Unlink and delete a device, invalidating its device token
      tags: *a2
  /devices/{id}/capabilities:
    patch:
      operationId: DeviceController_updateCapabilities
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDeviceCapabilitiesDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeviceResponseDto"
      summary: "Update device capabilities reported by the device itself
        (Authorization: Device <token>)"
      tags: *a2
  /devices/{id}/commands:
    post:
      operationId: CommandController_sendCommand
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommandDto"
      responses:
        "201":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponseDto"
      security:
        - bearer: []
      summary: Issue a remote command (RING, VIBRATE, MESSAGE, LOCK) to a protected
        device
      tags: &a3
        - Commands
    get:
      operationId: CommandController_listCommands
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
        - name: limit
          required: false
          in: query
          description: Number of items to return (1..50)
          schema:
            minimum: 1
            maximum: 50
            default: 20
            example: 20
            type: number
        - name: cursor
          required: false
          in: query
          description: Opaque cursor for pagination
          schema:
            example: eyJpZCI6IjEyMyJ9
            type: string
        - name: status
          required: false
          in: query
          description: Filter by command status
          schema:
            type: string
            enum:
              - PENDING
              - SENT
              - DELIVERED
              - EXECUTED
              - FAILED
              - EXPIRED
        - name: type
          required: false
          in: query
          description: Filter by command type
          schema:
            type: string
            enum:
              - RING
              - VIBRATE
              - MESSAGE
              - LOCK
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaginatedCommandsResponseDto"
      security:
        - bearer: []
      summary: List commands for a device with cursor pagination and optional filters
      tags: *a3
  /commands/{id}:
    get:
      operationId: CommandController_getCommand
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponseDto"
      security:
        - bearer: []
      summary: Get command status by ID
      tags: *a3
  /commands/{id}/ack:
    post:
      operationId: CommandController_ackCommand
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AckCommandDto"
      responses:
        "200":
          description: ""
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommandResponseDto"
      summary: "Acknowledge command delivery/execution from target device
        (Authorization: Device <token>)"
      tags: *a3
info:
  title: Guardian Mobile API
  description: Remote Anti-Theft Protection System for Android Phones
  version: 0.0.1
  contact: {}
tags:
  - name: Health
    description: System and database health monitoring
servers: []
components:
  schemas:
    HealthResponseDto:
      type: object
      properties:
        status:
          type: string
          example: ok
          enum:
            - ok
            - error
        service:
          type: string
          example: guardian-api
        version:
          type: string
          example: 0.0.1
        time:
          type: string
          example: 2026-10-05T23:50:00.000Z
        database:
          type: string
          example: up
          enum:
            - up
            - down
      required:
        - status
        - service
        - version
        - time
        - database
    RegisterDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: Secret123!
        displayName:
          type: string
          example: Alex Developer
      required:
        - email
        - password
        - displayName
    UserResponseDto:
      type: object
      properties:
        id:
          type: string
          example: 0c72e...
        email:
          type: string
          example: user@example.com
        displayName:
          type: string
          example: Alex Developer
      required:
        - id
        - email
        - displayName
    AuthResponseDto:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserResponseDto"
        accessToken:
          type: string
          example: eyJhbGciOi...
        refreshToken:
          type: string
          example: 8f7d92...
        expiresIn:
          type: number
          example: 900
      required:
        - user
        - accessToken
        - refreshToken
        - expiresIn
    LoginDto:
      type: object
      properties:
        email:
          type: string
          example: user@example.com
        password:
          type: string
          example: Secret123!
      required:
        - email
        - password
    RefreshDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: d3f6a2...
      required:
        - refreshToken
    RefreshResponseDto:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOi...
        refreshToken:
          type: string
          example: 8f7d92...
        expiresIn:
          type: number
          example: 900
      required:
        - accessToken
        - refreshToken
        - expiresIn
    LogoutDto:
      type: object
      properties:
        refreshToken:
          type: string
          example: d3f6a2...
      required:
        - refreshToken
    CurrentUserProfileDto:
      type: object
      properties:
        id:
          type: string
          example: 0c72e...
        email:
          type: string
          example: user@example.com
        displayName:
          type: string
          example: Alex Developer
        createdAt:
          format: date-time
          type: string
          example: 2026-10-06T00:00:00.000Z
      required:
        - id
        - email
        - displayName
        - createdAt
    LinkDeviceDto:
      type: object
      properties:
        installId:
          type: string
          example: install-uuid-12345
        name:
          type: string
          example: My Pixel 8 Pro
        platform:
          type: string
          example: android
          enum:
            - android
            - ios
        model:
          type: string
          example: Pixel 8 Pro
        osVersion:
          type: string
          example: Android 14
        appVersion:
          type: string
          example: 1.0.0
        mode:
          type: string
          enum:
            - PROTECTED
            - CONTROLLER
          example: PROTECTED
        fcmToken:
          type: string
          example: fcm-registration-token-sample
      required:
        - installId
        - name
        - platform
        - mode
    DeviceResponseDto:
      type: object
      properties:
        id:
          type: string
          example: c028a39a-7964-42b7-a36e-5264b3017a52
        ownerId:
          type: string
          example: u123...
        installId:
          type: string
          example: install-uuid-12345
        name:
          type: string
          example: My Pixel 8 Pro
        platform:
          type: string
          example: android
        model:
          type: object
          example: Pixel 8 Pro
        osVersion:
          type: object
          example: Android 14
        appVersion:
          type: object
          example: 1.0.0
        mode:
          type: string
          enum:
            - PROTECTED
            - CONTROLLER
          example: PROTECTED
        fcmToken:
          type: object
          example: fcm-token-sample
        adminEnabled:
          type: boolean
          example: false
        batteryLevel:
          type: object
          example: 85
        isCharging:
          type: object
          example: false
        lastSeenAt:
          type: object
          example: null
        isOnline:
          type: boolean
          example: false
        createdAt:
          type: string
          example: 2026-10-06T00:00:00.000Z
        updatedAt:
          type: string
          example: 2026-10-06T00:00:00.000Z
      required:
        - id
        - ownerId
        - installId
        - name
        - platform
        - mode
        - adminEnabled
        - isOnline
        - createdAt
        - updatedAt
    LinkDeviceResponseDto:
      type: object
      properties:
        device:
          $ref: "#/components/schemas/DeviceResponseDto"
        deviceToken:
          type: string
          example: d9b736b...256bit-token
      required:
        - device
        - deviceToken
    ListDevicesResponseDto:
      type: object
      properties:
        devices:
          type: array
          items:
            $ref: "#/components/schemas/DeviceResponseDto"
      required:
        - devices
    UpdateDeviceCapabilitiesDto:
      type: object
      properties:
        adminEnabled:
          type: boolean
          example: true
          description: Whether Device Admin privilege is active
      required:
        - adminEnabled
    UpdateDeviceDto:
      type: object
      properties:
        name:
          type: string
          example: Renamed Phone
        fcmToken:
          type: string
          example: fcm-new-token
    CreateCommandDto:
      type: object
      properties:
        type:
          type: string
          enum:
            - RING
            - VIBRATE
            - MESSAGE
            - LOCK
          example: RING
          description: Type of remote command (RING, VIBRATE, MESSAGE, LOCK)
        payload:
          type: object
          description: "Payload specific to the command type. RING: {durationSeconds:
            5..60}. VIBRATE: {durationSeconds: 1..30}. MESSAGE: {text: 1..200,
            contactPhone?: string}. LOCK: no payload allowed."
          example:
            durationSeconds: 30
      required:
        - type
    CommandResponseDto:
      type: object
      properties:
        id:
          type: string
          example: c123...
        deviceId:
          type: string
          example: d123...
        issuedById:
          type: string
          example: u123...
        type:
          type: string
          enum:
            - RING
            - VIBRATE
            - MESSAGE
            - LOCK
          example: RING
        payload:
          type: object
          example:
            durationSeconds: 30
        status:
          type: string
          enum:
            - PENDING
            - SENT
            - DELIVERED
            - EXECUTED
            - FAILED
            - EXPIRED
          example: SENT
        failureReason:
          type: object
          example: null
        createdAt:
          type: string
          example: 2026-10-06T12:00:00.000Z
        sentAt:
          type: object
          example: 2026-10-06T12:00:01.000Z
        deliveredAt:
          type: object
          example: null
        executedAt:
          type: object
          example: null
        expiresAt:
          type: string
          example: 2026-10-06T12:02:00.000Z
      required:
        - id
        - deviceId
        - issuedById
        - type
        - status
        - createdAt
        - expiresAt
    PaginatedCommandsResponseDto:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CommandResponseDto"
        nextCursor:
          type: object
          example: c123...
          nullable: true
          description: Opaque cursor for the next page, or null if no further pages
      required:
        - items
    AckCommandDto:
      type: object
      properties:
        status:
          type: string
          enum:
            - DELIVERED
            - EXECUTED
            - FAILED
          example: DELIVERED
        reason:
          type: string
          example: Sound played successfully
      required:
        - status